backend/pyproject.toml
[project]
name = "docqa-backend"
version = "0.1.0"
description = "DocQA Backend with FastAPI and Clerk Authentication"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "openai (>=1.92.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.2,<0.3)",
    "langchain-community (>=0.2,<0.3)",
    "langchain-openai (>=0.1,<0.2)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "clerk-backend-api (>=3.0.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
